Update README re 200 message limit / rejections
Due to the pagination of Mailman we'll get at most 200 messages at a
time for moderation. Supporting the pagination probably isn't a big
deal, but if you've got that many messages currently outstanding things
are going to be unwieldy anyway, so it's not a high priority.

Also mention that we don't allow message rejection yet (just
discarding).